refactor: extract spell audio manifest helpers into lib/spellAudio.ts
